+++ packages/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @acpus/web
 
+## 0.1.0-alpha.3
+
+### Patch Changes
+
+- Make WebUI shutdown idempotent under repeated Ctrl-C signals.
+
 ## 0.1.0-alpha.2
 
 ### Patch Changes